Removed three classes' MonoBehavior inheritance
I'm not as sure about this one, so Duane or anyone feel free to point
out my error.  Are we sure that Request, Response, and ProblemWorker
should inherit from MonoBehavior?  They do not seem to use any of the
superclass features, and they are not intended to be script components.
Removing the inheritance lets us instantiate them with the "new" keyword
without getting warnings from the MonoBehavior class.